simplify the server code
