+++ src/server/network/pending_client_session.cpp
@@ -2,7 +2,6 @@
 #include <common/network/net_command.h>
 
 #include "../world/server_world.h"
-#include "server.h"
 
 void PendingClientSession::sendHandshakeChallenge(u32 serversalt)
 {
